Port 3 more unit tests to Python 3
* barbican_manage.py: only decode bytes, not Unicode.
  six.string_types is str (Unicode) on Python 3.
* Replace long(1) with 1
* test_transport_keys_resource: decode HTTP body from UTF-8 to get
  Unicode.
* Replace ord(bytes[-1]) with ord(bytes[-1:]). On Python 3,
  bytes[int] returns an integer: use bytes[int:int] to get a
  substring.
* Remove following tests from tests-py3-blacklist.txt:

  - crypto.test_pkcs11
  - test_barbican_manage
  - test_transport_keys_resource

Partially implements: blueprint barbican-py3

+++ barbican/tests/cmd/test_barbican_manage.py
@@ -133,9 +133,9 @@ class TestBarbicanManage(TestBarbicanManageBase):
 
     @mock.patch('barbican.plugin.crypto.pkcs11.PKCS11')
     def test_hsm_gen_mkek(self, mock_pkcs11):
-        mock_pkcs11.return_value.get_session.return_value = long(1)
+        mock_pkcs11.return_value.get_session.return_value = 1
         mock_pkcs11.return_value.get_key_handle.return_value = None
-        mock_pkcs11.return_value.generate_key.return_value = long(0)
+        mock_pkcs11.return_value.generate_key.return_value = 0
         mock_genkey = mock_pkcs11.return_value.generate_key
         self._main_test_helper(
             ['barbican.cmd.barbican_manage', 'hsm', 'gen_mkek',
@@ -145,9 +145,9 @@ class TestBarbicanManage(TestBarbicanManageBase):
 
     @mock.patch('barbican.plugin.crypto.pkcs11.PKCS11')
     def test_hsm_gen_hmac(self, mock_pkcs11):
-        mock_pkcs11.return_value.get_session.return_value = long(1)
+        mock_pkcs11.return_value.get_session.return_value = 1
         mock_pkcs11.return_value.get_key_handle.return_value = None
-        mock_pkcs11.return_value.generate_key.return_value = long(0)
+        mock_pkcs11.return_value.generate_key.return_value = 0
         mock_genkey = mock_pkcs11.return_value.generate_key
         self._main_test_helper(
             ['barbican.cmd.barbican_manage', 'hsm', 'gen_hmac',
